buildPostNotificationPersonalisation
