+++ api/src/middlewares/services/EmailService/templates/notify/personalisationBuilder.postNotification.ts
@@ -1,0 +1,9 @@
+import { AnswersHashMap } from "../../../../../types/AnswersHashMap";
+
+export function buildPostNotificationPersonalisation(answers: AnswersHashMap) {
+  const post = answers["post"] as string;
+
+  return {
+    post,
+  };
+}
